fix: #763 - safer use of a distinct widget for attribute importance dialog
New files:
* `attribute_dialog.dart`: Vertical list of radio buttons used to set the importance of an attribute.
* `attribute_helper.dart`: Common values and methods for attribute display

Impacted files:
* `attribute_button.dart`: refactored by moving code to new files `attribute_dialog.dart` and `attribute_helper.dart`
* `user_preferences_page.dart`: minor refactoring
